# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Content, enquiry and admin API for Acquire 360 Ventures Ltd
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CompanyProfile {
  companyName: string;
  tagline: string;
  vision: string;
  mission: string;
  about: string;
  coreValues: string[];
  phone: string;
  email: string;
}

export type CompanyProfileInput = CompanyProfile;

export interface Service {
  id: number;
  title: string;
  description: string;
  category: string;
  /** @nullable */
  image?: string | null;
  featured: boolean;
}

export interface ServiceInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  category: string;
  /** @nullable */
  image?: string | null;
  featured: boolean;
}

export interface Project {
  id: number;
  title: string;
  client: string;
  description: string;
  category: string;
  completionDate: string;
  /** @nullable */
  image?: string | null;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  client: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  category: string;
  completionDate: string;
  /** @nullable */
  image?: string | null;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  category: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  specifications: string | null;
  /** @nullable */
  datasheet?: string | null;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  category: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  specifications?: string | null;
  /** @nullable */
  datasheet?: string | null;
}

export interface TeamMember {
  id: number;
  name: string;
  position: string;
  bio: string;
  /** @nullable */
  image?: string | null;
}

export interface TeamMemberInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  position: string;
  /** @minLength 1 */
  bio: string;
  /** @nullable */
  image?: string | null;
}

export interface Client {
  id: number;
  name: string;
  description: string;
  /** @nullable */
  logo?: string | null;
}

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  description: string;
  /** @nullable */
  logo?: string | null;
}

export interface ContentSummary {
  services: number;
  projects: number;
  products: number;
  clients: number;
  rfqs: number;
}

export interface RfqInput {
  /** @minLength 1 */
  companyName: string;
  /** @minLength 1 */
  contactPerson: string;
  email: string;
  /** @minLength 1 */
  phone: string;
  /** @minLength 1 */
  product: string;
  /** @minLength 1 */
  quantity: string;
  /** @minLength 1 */
  description: string;
  /** @nullable */
  file?: string | null;
}

export type Rfq = RfqInput & {
  id: number;
  createdAt: string;
  status: string;
};

export type RfqStatusInputStatus = typeof RfqStatusInputStatus[keyof typeof RfqStatusInputStatus];


export const RfqStatusInputStatus = {
  new: 'new',
  reviewing: 'reviewing',
  quoted: 'quoted',
  closed: 'closed',
} as const;

export interface RfqStatusInput {
  status: RfqStatusInputStatus;
}

export interface UploadResult {
  url: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthUser {
  id: number;
  name: string;
  email: string;
  role: string;
}

export type MediaAssetKind = typeof MediaAssetKind[keyof typeof MediaAssetKind];


export const MediaAssetKind = {
  image: 'image',
  document: 'document',
} as const;

export interface MediaAsset {
  id: number;
  filename: string;
  url: string;
  mimeType: string;
  size: number;
  kind: MediaAssetKind;
  createdAt: string;
}

export interface ErrorResponse {
  error: string;
}

export type ListProductsParams = {
search?: string;
category?: string;
};

export type UploadRfqAttachmentBody = {
  file: Blob;
};

export type UploadMediaBody = {
  file: Blob;
};