# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Content, enquiry and admin API for Acquire 360 Ventures Ltd
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  description: string;
  /** @minLength 1 */
  category: string;
  /** @nullable */
  image?: string | null;
  /** @nullable */
  specifications?: string | null;
  /** @nullable */
  datasheet?: string | null;
}